requirements.txt
vapi-python
python-dotenv
requests
